fix(clippy): the remaining lints, and set the ratchet to zero
Boxed the wide enum variants and PlanError out of its Result; named the complex
tuple types behind aliases; and a tail of small ones - is_multiple_of, integer
clamp, vec! initialisation, unused_mut, useless_conversion, unnecessary_cast,
hex grouping, a drop of a non-Drop value, a duplicated if arm, OnceCell to
OnceLock so KronosForecaster is Sync without an allow.

One variant is deliberately left unboxed: lfm2's Regime is several hundred
bytes wider than its sibling, but there is exactly one per model, held in a
field and never in a collection, so the padding is paid once while boxing would
put a pointer chase in front of every buffer binding in the per-layer loop.

Six allows in total across the whole pass, each carrying its reason. The one
worth reading is in the WGSL math oracle: `min(max(e, low), high)` there is not
a clumsy f32::clamp, it is the WGSL specification's own expansion of clamp, and
the two are different functions - f32::clamp propagates NaN where min/max
return the non-NaN operand. Swapping them would have quietly retuned the oracle
to whatever the implementation does.

The baseline goes to 0 in this commit, last, so the gate never reads zero while
warnings remain. From here the cost is paid one line at a time by whoever
writes the line, while they still have the context, rather than accumulating
into another 283-item burn-down.

+++ crates/qwen35moe/tests/sample_generate.rs
@@ -41,7 +41,7 @@ fn greedy_generation_is_deterministic() {
     let out1 = generate_kv(&model, &prompt, 6, 0.0, 0, 1.0, &[], &mut rng1);
     let out2 = generate_kv(&model, &prompt, 6, 0.0, 0, 1.0, &[], &mut rng2);
     assert_eq!(out1, out2, "greedy decoding must not depend on the rng seed");
-    assert!(out1.iter().all(|&t| (t as u32) < model.cfg.vocab), "generated a token outside the vocab");
+    assert!(out1.iter().all(|&t| t < model.cfg.vocab), "generated a token outside the vocab");
 }
 
 #[test]
